Removed variable, altered function
Removed the variable declaration "SLASH = "/" and instead directly compared the value "/" itself where needed.

Altered the IPCUTTER function to allow for IPGETTER to be removed. The logic has been changed to allow for a 0 or 1 flag to be passed when the IPCUTTER function is called to determine whether the IP address or netmask/CIDR is returned.